[IMP] timesheet_grid: Add timer for timesheets
TaskID: 2088586

Now, in the list, kanban and grid views in account.analytic.line model, we can update the
unit_amount field of a timesheet with a timer. This timer is only usable
if the timesheet is not validated and the encoding unit of this
timesheet is Hours.

To check if the encoding unit of timesheets if 'Hours', a new computed
field 'display_timer' is created into account.analytic.line model in
hr_timesheet module. this field is set to True is the encoding is 'Hours'.
With this field, we can easily hide and display the timer in different
views.

Also, the timer is only usable in 'My Timesheets' page. Because we want
that the user can use the timer only for his timesheets. Even if the
user have the access rights to see all timesheets, he will may only use
the timer for his own timesheets and nothing else.

In hr_timesheet, a timer.mixin has been created to use a generic timer
in the others models (like account.analytic.line and project.task).
In timesheet_grid module, the timer in account.analytic.line has changed
to use the timer mixin.

Remark : the condition that user can only use timer for his timesheets
hasn't used in kanban view because the kanban view is only displayed for
the 'My timesheets' page.

Timer in grid view :
------------------
The user can launched a timer for his timesheet of today.
When the user want to launch a timer for a task, a timesheet named
'Timesheet adjustment' is created and the timer for this
timesheet is launched. If today, the timesheet named 'Timesheet adjustment'
exists, then we use the timer of this timesheet.

The condition : 'Only one timer can be launched at the same time' is
managed for this grid view as for the other views (list and kanban)

The using of the timer in this view is possible only when the group by
is "project > task", Why ? Because with another group by, the data
retrieve in the js is different and difficult to have a generic use for
the timer. For example, the user change the group by and group just by
project, we lose the information about the tasks and we can't create a
timesheet for a task in the project.

Add style to timer button

Some improvements for grid view in Timesheets :
---------------------------------------------

Add 'Today' button in grid view (for all grid view with range of date)

Add stylesheet for timesheet_grid view to give fix width for the timer.
Thanks to this, when the user want to start the timer, the width column with
the task name and timer doesn't change every time.