[FIX] sale_amazon: add company checks to amazon_account
Prior to this commit, it was possible to assign to an amazon account a salesperson, sales team or stock location with a different company than that of the amazon account, which later resulted in an error (depending of the field concerned).
This is fixed by adding company_checks to all relational fields in amazon_account.

task-2081012

X-original-commit: 96f096db7394ebd2d16bc2db8dc31b877fdc45f8