[FIX] (sale_)account_taxcloud: auto-generate taxes in active state
As of #13a39c3 TaxCloud taxes were auto-generated in inactive state to avoid clustering the views and hiding the records from the user. Because the computation of the totals of sales orders and invoices does not take archived taxes into account, the totals were wrong.

The fix that this commit introduces for this issue is to auto-generate TaxCloud taxes in active state again and toggle the active field of archived taxes that are matched when searching for an already existing tax. As the user won't browse the taxes by himself in most flows, because TaxCloud is responsible for the tax assignation and declaration, the clustering of the views is not a real issue.

task-2081012

closes odoo/enterprise#6456

X-original-commit: 0356933bee991a56279864421ae62e07defd025b
Signed-off-by: Antoine Vandevenne (anv) <AntoineVDV@users.noreply.github.com>