[FIX] sale_amazon_taxcloud: request TaxCloud to fetch taxes
Prior to this commit, TaxCloud never applied the right taxes on sales orders synchronized from Amazon as they were directly created either in the 'sale' or 'done' state, while TaxCloud is called in the 'action_confirm()' method.

This commit adds a call to sale_account_taxcloud to trigger the fetching of the right taxes after the synchronization of an order from Amazon.

task-2081012

closes odoo/enterprise#6556

X-original-commit: 8dc9b6e7a3e7710133653e8ed2d6feeade40f826
Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>