[MOV] hr_work_entry_*: Remove work_entry generation from payroll
PURPOSE
========

It isn't possible to use work entries without Payroll. And it isn't
possible to use payroll without Time Off.

Before this commit:
1. If we install work_entry application, we have no menu neither views.
And if we add a view for this model, no work entries are created.

2. If we use payroll application, Time Off application is also installed.

After this commit:
We have 4 different applications:
1. hr_work_entry: Basic models and views, no menu. No modification
is made in this module.

2. hr_work_entry_contract: Add menu payroll with only work entries.
Add work entry generation mechanism based on contract and calendars.
Time off from resource.calendar.leaves are also taken in account.

3. hr_work_entry_holidays: Depends of hr_work_entry_contract, links
work entries with hr_leave. Management of conflicts due to leaves not
approved.

4. hr_payroll: Depends of hr_work_entry_contract. When we generate a
payslip, work entries are generated for this employee. When we manually
adapt work entry lines (number of work hours), amounts are re-computed.

closes odoo/enterprise#6613

Taskid: 2116977
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>