[IMP] hr_appraisal: Allow manager to request appraisal
Since c9ca376, only user in group_hr_user can request an appraisal
for their subordinates through the employee form views.
The reason is the button is only displayed on `hr.employee` forms.

A manager (`parent_id`) of an employee should be able to request appraisals.
This commit adds the button "Request Appraisal" to the public employee form.
The button should appear only on subordinates employees.

This commit removes the default value fallback to the private address of
the employee because the manager does not necessarily have access rights
to read it.

Task 2028794