[IMP] sale_amazon: update default offers with newly matched product
Prior to this commit, if a synchronized SKU was found in an offer, the associated product was set to the sales order line in all cases. This lead to situations where the user would later set the corresponding SKU as internal reference of another product and expect the matching to happen but, as the offer was never updated, the default product was still linked to this SKU.

This commit makes the search for a matching offer go through all products again to find one with an internal reference matching the current SKU if the product currently associated with the offer is the default product.

task-2081012

X-original-commit: c46823591637133fa98346de2790b20346934c92