[IMP] hr_appraisal: Create activity when requesting appraisal
When an employee requests an appraisal, an activity should be created for his
manager: "Confirm and send appraisal"

In the same way, if a manager requests an appraisal for one of his employee,
the employee should recieve an activity.

The activity should not be notified to the user because he already recieves
an email.

Task 2028794