[IMP] iot: Easier detection of new devices
Until now, to detect new types of devices on the IoT Box, we had to
modify driver.py. If partners wanted to connect a new type of device,
they had to modify directly the files on the Box, meaning that after
each new build they had to re-apply all their modifications.

We remove the loops that detected the connected devices and replace
them by `Interfaces`. All interfaces will be loaded from the connected
Odoo instance with the corresponding drivers.

closes odoo/enterprise#6552

Taskid: 2082282
Related: odoo/odoo#39759
Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>