[FIX] sale_amazon: display gift messages as notes on orders
As Amazon dropped the possibility (at least on the amazon.com marketplace) to choose paid gift wrappings for order items, causing gift messages to not display on sales orders anymore.

This commit makes gift messages being displayed as notes on orders, whether the gift wrapping is charged or not.

This allows to match a product with gift wrapping charges if any or to only display the gift message in the other case.

task-2081012

X-original-commit: a10b392d44d618e1a0b9c8240d64bc23e9a418a5