[ADD] sale_coupon_taxcloud_delivery: discounts, delivery, and taxcloud
When sale_coupon_delivery is installed, the delivery lines are not
discountable anymore. Meaning that with only sale_coupon and delivery,
a percentage discount on order would also apply to the delivery,
whereas this is false once sale_coupon_delivery is installed.
It also adds another type of reward, the free delivery, which conversely
applies only on delivery lines.

opw 2079677

X-original-commit: 010926e28c08226531b21466ee092fe503aff497