[REF] iot: Make IotRealTimeValue more generic
IotRealTimeValue was implemented in the `iot` module but could only be
used in `quality_iot` because of the check of the `test_type`.
We make it more generic so that we can use it in other modules and
override it in `quality_iot`.
We also move `getDeviceInfo` to the Mixin to avoid duplicating code in
`iot_picture`.