[IMP] hr_appraisal: Allow multiple recipients when requesting appraisal
Purpose
=======
Currently, the appraisal request mail is only sent to one person.
The mail is sent to the employee manager if the requester is the employee himself,
or to the employee if someone else requested the appraisal.

But if a DRH requests an appraisal, both the employee and the manager should
receive the email.

Specification
=============
Allow to set multiple recipients: One2many => Many2many
By default the value should be:
- the employee if it's the manager who's requesting
- the manager if it's the employee who's requesting
- both if it's someone else

If both the employee and the manager are recipient, the default template should
be the one for managers. It's up to the current user to customize it to fit the
situation.

Task 2028794