[FIX] account_reports: activity update
- Create a DB without demo data
- Configure a mail server
- Set a country to the company, e.g. 'United States'
- Install 'Accounting'

A next activity email is sent about 'Tax Return'.

This email is confusing for new users. Moreover, on the SaaS, this
happens before the first connection of the user. In this case,
`web.base.url` is not set yet and the email contains a link to
`localhost`.

There is no need to send an email for this activity.

opw-2115458

closes odoo/enterprise#6499

X-original-commit: 582859b9844a6ec783b1a458cbf5ef1e981253d3
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>