[FIX] account_reports: indentation of auto-hierarchy
It was really hard to read the financial reports since the indentation
was chaotic: all accounts were set on the second indent level even if
the groups were indented further, and the values of debit and credit
were kinda centered instead of right aligned.
Because the alignement has changed, we also had to add a border around
debit and credit so that we can see what the 'super-columns' are
refering to.

closes odoo/enterprise#6564

X-original-commit: 7e37d1b863b153ec2392f39bbf570a43325efcc3
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>