[FIX] helpdesk_account: replace onchange domain by static domain
This may not be the idealest version. There are basically two ways to
replace a non-trivial dynamic domain by a static one:

First is to make the condition into a computed many2many which serves
as a records whitelist, this is pretty readable and can reuse the
onchange logic, it can also be fairly easily overridden (override the
compute method), however it doesn't really show anything and the
scaling can be debatable.

Second is to encode the conditional bits by AND-ing or OR-ing
TRUE/FALSE leaves with other segments e.g.:

    if condition: domain.append(predicate)

becomes

    [OR (not condition = 1) predicate]

this way the predicate is ignored IFF `condition` is false. However
this is somewhat tricky to read especially for more complex
domain (let alone if the original has both `if` and `else`
branches). In the latter case it's also possible to use conditional
expressions:

    [pred0 if condition else pred1]

however this has the drawback of the entire expression being opaque,
and thus the ability to analyze and report on the segments being
limited.

Here here the conditional *sequence* (if/elif/else) makes the 2nd
option basically unusable, so rely on a computed field instead.

Task 2115472

closes odoo/enterprise#7087

Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>