[FIX] account_reports: fix multicompany tax reports
Be in multi-company. Download an accounting tax report.
If have only one company selected (who is not the default company),
the report will display the information of the default company.

It happens because the tax report doesn't have the multi_company option
and there is now allowed_company_ids in the context.

We set them in the context using the `cids` cookies so that the
correct report is downloaded.

This commit impacts all the reports who doesn't have the multi_company
option.

Related to task 2091557
OPW-2091214

closes odoo/enterprise#6542

X-original-commit: ba093a4728eae5280912ba71d4f0ccc09cac7945
Signed-off-by: Jason Van Malder <jasonvanmalder@users.noreply.github.com>
Co-authored-by: len-odoo <len@odoo.com>
Co-authored-by: rgo-odoo <rgo@odoo.com>