[FIX] helpdesk_[sale_]timesheet: replace & fix onchange domain by static
Initial instinct was to simplify the domain and e.g. always filter
tasks on company, but after more thinking it makes sense that the
company filter be skipped if a project is selected: the company of the
project & that of the ticket could diverge if the project was moved
between teams or something, and in that case it makes sense that we'd
allow selecting other tasks of that project.

With the company filter, we'd not be able to select any task from the
project.

Use a computed m2m as filter, so it can be overridden in
helpdesk_sale_timesheet instead of the two modules playing tug-of-what
or whatever.

Also remove the task_id reset in helpdesk_sale_timesheet, unsetting
the task when we change the partner_id but use_helpdesk_sale_timesheet
is not set is... odd? The parent module already has a task_id reset
when we change to an other project.

Task 2115472

closes odoo/enterprise#7084

Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>