[FIX] account_intrastat: ec sales list report
Before this commit, the value in the EC sales list report was always 0.
It was adding the balance of the move lines, and each move has two move
lines one with an amount in the credit and one with an amount in the
debit.

Now, the value is the sum of the invoices (moves)

opw-2081572

closes odoo/enterprise#6557

X-original-commit: 1530d31dd3c25ffd520819d90e4d904e0a22b865
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>