[FIX] l10n_mx_edi: Use the correct expedition address on payment
Recently the SAT adds a new validation, to ensure that emission date of
a CFDI is correct according to the emission address of a CFDI.

Then, now validates that the date is correct based on the
``LugarExpedicion`` attribute on the CFDI.

Then, is necessary allow use the branch address when is emitted a
payment complement (How in the invoice). For this:

1. Was improved the journal view to allow set the branch address.
(Before only in sale journals are possible)
2. Send the issued address in the CFDI values
3. In the attribute for ``LugarExpedicion`` check if the issued address
is set, and get from that the ZIP, else from the company address.

This is the same process that in the CFDI for invoices

closes odoo/enterprise#6644

X-original-commit: 4bba87c64533e55998a53a28a4588cb285d67a77
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>