[FIX] l10n_mx_edi: Portal Invoice QR code layout and String formatting.
In version 12.0 the portal report assets were stored in the
`assets_common` bundle, in version 13.0 the report assets are properly
set in it's own bundle `report_assets_common`.

This commit moves the assets for e-invoicing to `report_assets_common`
bundle so the QR Code and the Original String are properly rendered
in all reports.

closes odoo/enterprise#6555

X-original-commit: 6d2ca21114bdb3465307a959d93625f80159a2d1
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>