[FIX] web_map: alert box should not be allow to grab the map
before this commit: you can grab the map on alert box, so you can
perform all the operations of map by clicking or hovering on the alert
box above the map.

after this commit: we are preventing all the operations like grabbing or
zooming on the alert box, also changed the alert box from a warning to
info message.

task-2093889

closes odoo/enterprise#6535

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Co-authored-by: Mohammed Shekha <msh@odoo.com>