[FIX] planning: Timezone issue when creating a slot
Steps to reproduce the bug:

- Go to planning app
- Set the gantt view
- Create a slot for today with planning template between 8 AM - 2 PM

Bug:

The hours displayed were not in the right timezone

Technically, start_time from self is in UTC (not in the user timezone), so when changing
the hours and minutes from the wizard, the datetime must be converted
in the timezone user. Otherwise the hours and minutes will never be
displayed as the values set in the wizard.

Example: start_time = 06/11/2019 at 8:00 am in UTC +10 must be saved as 05/11/2019 10:00 pm in UTC
Before the fix, 06/11/2019 was kept and just the hours and minutes were changed.

opw:2091602

closes odoo/enterprise#6575

X-original-commit: 83881baa8d7fc88ed38ba25bec56310f0098cbb0
Signed-off-by: Simon Goffin (sig) <sig@openerp.com>