[FIX] account_reports: improve layout of email report
Install account_reports. From Accounting>Customers>Follow-up Reports
select a followup and click on "Send Email".

Aoctivating debug mode, and looking into the email queue is possible to
see that the layout of the email is too much clobbered.

The class-dependant style is applied by the transcoder
https://github.com/odoo/odoo/blob/12.0/addons/web_editor/static/src/js/editor/transcoder.js
but is not the case when sending the email, so the directive
"width=100%" allow to achive an acceptable view.

opw-2088622

closes odoo/enterprise#6595

X-original-commit: 0eddfa5eb9c197283597905852413911c50ade28
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>