[IMP] sale_account_taxcloud: ensure 'all at once' invoicing
Ensure that any TaxCloud order that has discounts is invoiced in one go.
Indeed, since the tax computation of discount lines with Taxcloud
requires that any negative amount of a coupon line be deduced from the
lines it originated from, these cannot be invoiced separately as it be
incoherent with what was computed on the order.

We could allow simple cases to be partially invoiced,
e.g. only one coupon that is a line specific discount.
However we need to make sure the matching lines are invoiced together,
and this does not solve the general case
(e.g. any global discount require total invoicing in one go).

Co-authored by: Nans Lefebvre <len@odoo.com>

X-original-commit: 3a5faf90fe25bbe7b229aa6be153607f02553169