[IMP] helpdesk: Correctly link reversal lines to tickets
In account, we added new move reversal lines in the wizard model,
those new_move_ids are usefull for the override in helpdesk.

Currently, we return a form or a list view according to the
number of generated reversed moves.

Now, we link the generated moves on the wizard, so that we can
link the invoice lines correctly to the helpdesk tickets. Now,
the invoice lines are correctly linked only if there is 1 reversed
line.

id=2017554

closes odoo/enterprise#5986

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>