[IMP] timesheet_grid: improve validation process and add group_expands
TaskID: 2088586

The validation of timesheets use only a boolean field named 'validated'
in account.analytic.line model. Before the validation use the validation
date for an employee but this field has been removed because, the
validation valided all timesheets of an employee before the validation
date, even the timesheets in others projects where the approver can't
see these projects.

The approver can validate timesheets only if it's manager of employee
or timesheet Responsible.

The title of views in 'To Validate' menu has been renamed.

Adding a group_expand into project_id and employee_id fields in
account.analytic.line, it allows to display either the projects where
the user timesheeted in the past 30 days or the employees who
timesheeted in the past 30 days in grid views.
Modify the definition of timesheet.validation and
timesheet.validation.line models.

The wizard for the validation timesheets contains records group by
employee > project > task. Now the user can validate the timesheets by
project and by task (if task exists).

Warning, the method to group by the timesheets could be improve to gain
more performance (in action_validate_timesheet method into
account.analytic.line model).

Before this commit, the 'To validate' menu display either the previous
week or the previous month. The problem is if a non-validated
timesheet is before the previous week or month, than it can be never
validated by the approver or the manager.
To revolve this problem, in this commit, we search the oldest
non-validated timesheet and display the grid view that contains this
timesheet. Thus, the approver can be sure that he has approved all
timesheets before the range date of grid view.

Add domain of the filter in search for records added with the group_expand

Testing :
-------
For every access rights, we check the read/create/update/delete
operations.