[FIX] l10n_nl_reports: fix report balance
Fine-tuning of 911b88da2516f.
While it makes sense to subtract a cost (as is done on other reports)
on this report the cost lines have all balance = -sum.balance.
It turns out that this commit was written for a database on which
b46f4bdc27de569e had not been applied, and for which the fix corrected
the report.
Since this report set cost lines as negative, we keep it that way.

opw 2115991

closes odoo/enterprise#6953

X-original-commit: 42389e40d63453ba295dcbebba2319fad6942a10
Related: odoo/odoo#41071
Signed-off-by: Nans Lefebvre (len) <len@odoo.com>