[FIX] social: fix wrong emoji click event in ``stream.post comments``
Bug
===
In the feed view, if you click on an emoji (which is inside of a comment) there will be a traceback

Fix
===
In the file ``stream_post_comments`` we bind ``click .o_mail_emoji`` to the emoji mixin method.
This method will insert the emoji in the text.
But, this event is also triggered if the user click on an emoji in read only mode.
So, we need to change the target of the event (and to select only emojis which are in the dropdown menu).

Task #2121750

closes odoo/enterprise#6867

X-original-commit: 5109fde1bfdd9c5b18e256f1054a8ad894e84367
Signed-off-by: std-odoo <std-odoo@users.noreply.github.com>