[IMP] sale_coupon_taxcloud: test coherency between price_taxcloud and order
This test defines a set of programs.
For each subset of the powerset, it applies that subset to an order.
Then we _apply_discount_on_lines to obtain the price_taxcloud that will
be sent. To be coherent, the sum of these prices (by the quantities)
should match the total of the order (which uses the prices minus the
discounts).

X-original-commit: 44ec1e577c29706175a99ec9c460766595284e4f