[FIX] currency_rate_live: Current date used to get rate from banxico
Now Banxico provides the rate that will be used in the CFDIs every day,
and the date could be used as parameter to get the specific date.

With this, Banxico provides the serie ``SF60653``, and that serie
returns the rate that must be used in the day that is downloaded.

Some series are not updated early, but yes the USD, that is the more
common currency used in Mexico after that MXN, for this, update the rate
from today and yesterday.

closes odoo/enterprise#6661

X-original-commit: dcf15ab5cacab2d71ec0d62a82ce91675f5f99eb
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>