[FIX] account: payment reference exact match without spaces
Cherry pick of https://github.com/odoo/odoo/commit/7347ff9a9badcb04502b214280009fe967f3ebec

The spaces should not be important in a structured reference.
Also, we should not have to include the second batch as that batch
(previously the first) was intended to match the payment reference.
By including it, we may have too much false positives proposed in the
widget and even worse: auto reconciled.

closes odoo/enterprise#6645

Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>