[FIX] account_asset: set Fixed Asset Account at copy
- Go to Account > Configuration > Deferred Revenue Models
- Duplicate the demo record
- Save

Impossible to save because Fixed Asset Account is mandatory but not set.
Moreover, it is hidden, so there is no way to set it.

Since `original_move_line_ids` is empty, the computation of the field
result in a `False` value.

We force the value when copying in order to keep it.

opw-2122998

closes odoo/enterprise#6678

X-original-commit: eac058d50dd8750d325457f869d40f1a244cdcd3
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>