[FIX] currency_rate_live: Banxico provider: properly handle bad responses from web service
Before that, an error message was shown to the user when Banxico didn't answer, saying that the rate of the main currency was not supported by Banxico (whatever the currency, it happened with EUR on runbot: http://runbot.odoo.com/runbot/build/683935). This was due to the way the values returned by Banxico were managed if this answer meant no information could be retrieved. v12 API requires to return False in this case, not a (probably vert) partial mapping of exchange rates.

closes odoo/enterprise#6759

X-original-commit: e1b5e36558a5a1796be5f38ccf81f5826c6d7f69
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>