[FIX] account_taxcloud: forbid invoice cancellation
At invoice confirmation, the invoice is sent to TaxCloud;
it is then accounted for in taxes that the seller owe to the state.
To cancel it, a credit note should be issued to the end customer;
the taxes are thus correctly cancelled.
This flow should always be the preferred one.

closes odoo/enterprise#7313

X-original-commit: 4752d3c6465494f874731ef941e57c110ef76f9c
Signed-off-by: Nans Lefebvre (len) <len@odoo.com>