[FIX] stock_barcode: detect validation properly
In the barcode app, select a picking and process only part of the
reserved units. When validating the transfer, you get the wizard asking
to create a back order/do not create a back order. After clicking on any
of those buttons, the picking should be set as done and we should be
redirected to the Kanban with the different operations. This isn't the
case at the moment.

After 0bf9217f01d8cf2b0fe1ed533c4afde9286bc909 the structure of the
'infos' variable has been changed, however the behavior is the same:

* On Apply (green button) infos is undefined
* On Cancel (red button) infos is defined, contains special: true
* On window close (x button) infos is undefined

This change however affect how the "not special" case should be handled,
because 'infos && !info.special' does not pass the check when infos is
undefined

opw-2124191

closes odoo/enterprise#6722

X-original-commit: a4129bd30cdf9316d9028f28920a27a981e5c8ea
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>