[IMP] pos_loyalty: improve usuability of the loyalty programs
Changed the view in the back end and removed the point per product and currency earned.
Changed the way we manage rules. Instead of adding point per product and currency, we give a a fixed amount if the rules is valid or not.
If the rules is valid, we can choose to add point one or multiple time this amount.
We can also choose to validated rules only if the previous one are valid.

The reward points consumption is changed too.
When we use percentage:
if the reward is on the order, we remove a fix amount of points.
If the reward is on the order line, we remove the amount of product * nb of points.
When we use fixed price, we just remove the fixed amount.
We can also set a rule on the rewards order and if the previous is not valid, we can't receive the next.

closes odoo/enterprise#6736

Task-id: 2069831
Related: odoo/odoo#41658
Related: odoo/upgrade#984
Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>