[IMP] account_predictive_bills: predict taxes
Task 2092094
Before, account_predictive bills only predicted the account and the
product. When the product is detected, the taxes are taken from the
product but there are cases where you fdon't have the product but still
want to detect the correct taxes.
This commit adds the taxes to the line if there is no product detected.

closes odoo/enterprise#6757

Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>