[FIX] delivery_easypost: UPS delivery with multiple packages
Usecase to reproduce:
- Create a delivery order with easypost UPS as carrier
- Use multiple packages
- Validate the picking

It raise a Usererror returned from easypost response. However the buy
order is created on easypost.com

It happens because other carrier from easypost during an order return
the rate for the order in the first shipment and following shipments have
their rate empty. UPS do the same process but it adds a message in shipments
with empty rate explaining that the rate is in the first shipment.
However when easypost request catch a response with a messages inside,
it consider it has an error message and return it to the customer.

In this case the error message is not a real error and there is no flag
in the response that allow us to know if it's a real error or not thus
for this case we remove it from the response and log it as a
message_post on the picking.

opw-2081066

closes odoo/enterprise#6997

X-original-commit: 1c0e6326e4b0fad1436413b03efd8edd0a6f3b23
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>