[IMP] sale_timesheet_enterprise: recompute the sale.order.line
TaskID: 2088586

When we change the configuration in tiemsheet to invoice either all
timesheets or only approved timesheets, we recompute the qty_delivered,
qty_to_invoice and invoice status in sale order line service of each
sale order with the state 'sale' or 'done'. For qty_to_invoice, we need
to check if the product has a service policy set on 'Timesheet on task',
if yes, then we update the qty_to_invoice of sale order line.
With this recompute, the sale order created before the
changement of the configuration can be affected by the changement.
Before this commit, it wasn't the case.

Add a file to demo data, because we need to validate the timesheets
created in sale_timesheet module and where these timesheets are
before the last week. Then, in the 'to validate' pages, the oldest
non-validated timesheet is last week and not 3 months in the past.