[FIX] account_winbooks_import
OPW 2126259
Winbooks uses unique ids for things that should be different.
* The journal entries numbers: we do it the odoo way. Instead of
importing the numbers from Winbooks and put it as numbers in Odoo, we
recompute them. We keep the numbers as information in the `ref` field.
* The matching numbers: we give a different number for each account when
there is a single number in Winbooks that matches lines in multiple
accounts.

closes odoo/enterprise#6975

X-original-commit: b5c99c2fb4b7d34c3565f061adfc27631826ec83
Related: odoo/odoo#41138
Signed-off-by: wan <william-andre@users.noreply.github.com>