[FIX] delivery_ups: call production server not only test...
Usecase is rather simple do a delivery order with UPS in a production
environement. The return label and object on UPS created are samples.

It's due to [1] that update suds to zeep and
'''
security['location'] = '%s%s' % (self.endurl, api)
'''
just do nothing so the processShipment call will take the wsdl soap
address that is the test server url.

Fix it by creating a service with the correct production url or test
url.

[1] commit 8744f072168d653a4dfa9d6e0cf62c80321d2688

opw-2125049

closes odoo/enterprise#6805

X-original-commit: 03081226315b9ec9a4d0bc9759b3b3b3a95e8a43
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>