[IMP] approvals: replace item by product
In approval request, replace the `item` field (string) by a one2many
field: `product_line_ids`.
The 'approval.product.line' model makes the link between the approval
request and the products.
If quantity is optional or required for this approval category, the
quantity will now be written on the product lines, not on the request
directly.
Also, adds stock in approvals' dependencies.

task-2040710