[IMP] mrp_workorder: add extra components
This commit allows the manufacturing user to add additional component or
additionnal byproduct during a production directly form the workorder
tablet view.

The new product is added via the burger menu of the tablet view. The
process is the following :
 - The 'add component' (add byproduct) open a little wizard to choose the
   new product, the quantity and eventually the uom.

 - Closing the wizard will create a new wo line, a new quality check and
   set the workorder on it.
 - The record production will create a stock_move (raw or finished) for
   the additional workorder lines

The 'add component' feature is available once the bom has the flexible
consumption setting activated.
The 'add byproduct' feature is available once the by_product group is
activated.

Task : 2082448

closes odoo/enterprise#6366

Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>