[IMP] mail: Unsychronize chat windows between mobile and desktop
Currently when you open, for example, 5 chat windows on the desktop,
each of those windows will be opened on mobile as well.
The big issue with that is that I have to shut down all windows before
starting to work on the application.

To make things simple, you have to explicitly open chat window to
make it appears on mobile. Synchronization is blocked in both direction.
We don't update the chat window state on the server and we no longer
consider notifications (bus) about it.

We had to add this logic in community repository for ease, but as
this is an enterprise feature, tests were added here.

Task ID: 1891952

closes odoo/enterprise#6943

Related: odoo/odoo#41052
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>