[FIX] ocn_client: avoid traceback on server-side issue
Before this commit, the client would throw an exception when the OCN
server was down, more specifically on the creation (save) of a
res.config.setting and on method `_notify_recipients` from `mail.message`
thus preventing the sending of notifications altogether.

This commit alters that behaviour by catching the exception and logs it
instead, allowing the notification flow to continue.

closes odoo/enterprise#6303

closes odoo/enterprise#6992

Original-signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
X-original-commit: d240c4aab2b695a1a3623041ff90cd81b037a40f
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>