[FIX] account_sepa: force account in register payment
- Activate SEPA Credit Transfer for outbound payments for 'Bank'
- Create and validate 2 vendor bils for the same supplier
- Select both bills, then Action > Register Payment
- Choose SEPA Credit Transfer

You are allowed to validate, but the payment will be created without a
bank account set. Later in the process, the user will be blocked since
it won't be possible to set the bank account anymore.

The bank account taken is from the first invoice of the list:

https://github.com/odoo/odoo/blob/4eb47d335c5c87c9cc39a504e67b25d026b3c422/addons/account/models/account_payment.py#L795

We force the user to set the bank account on all invoices.

opw-2119374

closes odoo/enterprise#6823

X-original-commit: 0599a1836b4a326ba48edc24f2d062ec6b47bc16
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>