[FIX] web_gantt: avoid char conversion in date
Activate and set as language Arabic(Syria). Go to Payroll>Work
Entries>Conflict

Traceback will popup: date has been converted to arabic characters and
database does not recognize the date object. This is because moment.js
change the date according to the locale of the system, so the locale
need to be manually set to 'en' just to evaluate the datetime.

opw-2122916

closes odoo/enterprise#6844

X-original-commit: 8b87bf130704fa771eb45c58824d1f14e708e2c5
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>