[FIX] account_batch_payment: should depend on account_accountant
account_batch_payment overrides / updates the reconciliation widget.

That widget was moved from account to account_accountant by #6032,
however the dependency was not updated, so when installing
account_batch_payment or one of its dependents (e.g. account_sepa or
l10n_au_aba) the installation blows up and fails as the reconciliation
widget is not found.

closes odoo/enterprise#6838

Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>