[FIX] sale_ebay: fix product image upload
Revert 495875153b and actually solve the problem it was trying to solve.

Ebay introduced a breaking change: instead of just displaying a warning
when uploading a too small image, it started to return an error.
The current condition is that images should be of at least 500 pixels
on the longest side.

Instead of filtering images based on this condition, it broke the image
uploading feature by only checking the attachment on the res_field
'image', and started to ignore all other attachments.

[IMP] sale_ebay: only upload images following ebay guidelines

There are more conditions to check for an image to be valid for ebay.
We don't implement the check on the size, which is partially redundant
with the maximal image size.
Heavy lossless images might go above the 7mb limit for images hosted
on eBay, but to compute it accurately we need to save the file on disk,
so we'll just skip it.

opw 2126399

closes odoo/enterprise#6911

X-original-commit: 903e8f1c62d6d6a8542560418c3471a2b0b2490a
Signed-off-by: Nans Lefebvre (len) <len@odoo.com>