[IMP] fleet: Add analytic account and mobility card
Create analytic account from fleet (and link it to vehicle) based
on the plate and the fiscal deduction rate (if l10n_hr_payroll_fleet
is installed).

Add on employee a Mobility Card number. This number is set on vehicle
when we put a partner linked to employee on a vehicle.

closes odoo/enterprise#6865

Taskid: 2127641
Related: odoo/odoo#40795
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>