[IMP] event_barcode: improve barcode usability
PURPOSE

Before this task, you sometimes had to scan the same attendee
multiple times before finding his registration. This is because
the barcode interface searched only for a specific event.

SPECIFICATIONS

We removed the filter on the specific event when manually selecting
an attendee in case you have multiple events at a single registration
desk. People will always have to type the attendee's name anyway so
it doesn't change much. We replaced that filter with a filter on the
"ongoing events" to get a clean list of the attendees.

LINKS

Task ID : 2093336
PR : #7622