[FIX] iot: Six terminal loses configuration
If the shift was open during a restart the MPD server or a reboot of
the terminal, the shift was automatically re-opened but the
configuration (language) was lost.

Furthermore, we used to check the state before opening/closing, which
was useless because EFT_Open and EFT_Close already do it.

closes odoo/enterprise#6937

X-original-commit: 54e157914e7c6a60a6d1fd50181dd745913adad6
Signed-off-by: Antoine Prieëls <aprieels@users.noreply.github.com>